# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Open Archiver API",
    "version": "1.0.0",
    "description": "REST API for Open Archiver — an open-source email archiving platform. All authenticated endpoints require a Bearer JWT token obtained from `POST /v1/auth/login`, or an API key passed as a Bearer token.",
    "license": {
      "name": "SEE LICENSE IN LICENSE",
      "url": "https://github.com/LogicLabs-OU/OpenArchiver/blob/main/LICENSE"
    },
    "contact": {
      "name": "Open Archiver",
      "url": "https://openarchiver.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT obtained from `POST /v1/auth/login`. Pass as `Authorization: Bearer <token>`."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY",
        "description": "API key generated via `POST /v1/api-keys`. Pass as `X-API-KEY: <key>`."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication is required or the token is invalid/expired.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorMessage"
            },
            "example": {
              "message": "Unauthorized"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The authenticated user does not have permission to perform this action.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorMessage"
            },
            "example": {
              "message": "Forbidden"
            }
          }
        }
      },
      "NotFound": {
        "description": "The requested resource was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorMessage"
            },
            "example": {
              "message": "Not found"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Human-readable error description.",
            "example": "An error occurred."
          }
        },
        "required": [
          "message"
        ]
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Operation completed successfully."
          }
        },
        "required": [
          "message"
        ]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clx1y2z3a0000b4d2"
          },
          "first_name": {
            "type": "string",
            "nullable": true,
            "example": "Jane"
          },
          "last_name": {
            "type": "string",
            "nullable": true,
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "jane.doe@example.com"
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "createdAt"
        ]
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clx1y2z3a0000b4d2"
          },
          "slug": {
            "type": "string",
            "nullable": true,
            "example": "predefined_super_admin"
          },
          "name": {
            "type": "string",
            "example": "Super Admin"
          },
          "policies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CaslPolicy"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "policies",
          "createdAt",
          "updatedAt"
        ]
      },
      "CaslPolicy": {
        "type": "object",
        "description": "An CASL-style permission policy statement. `action` and `subject` can be strings or arrays of strings. `conditions` optionally restricts access to specific resource attributes.",
        "properties": {
          "action": {
            "oneOf": [
              {
                "type": "string",
                "example": "read"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "read",
                  "search"
                ]
              }
            ]
          },
          "subject": {
            "oneOf": [
              {
                "type": "string",
                "example": "archive"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "archive",
                  "ingestion"
                ]
              }
            ]
          },
          "conditions": {
            "type": "object",
            "description": "Optional attribute-level conditions. Supports `${user.id}` interpolation.",
            "example": {
              "userId": "${user.id}"
            }
          }
        },
        "required": [
          "action",
          "subject"
        ]
      },
...
    }
  },
  "paths": {
...
    "/v1/users": {
      "get": {
        "summary": "List all users",
        "description": "Returns all user accounts in the system. Requires `read:users` permission.",
        "operationId": "getUsers",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of users.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create a user",
        "description": "Creates a new user account and optionally assigns a role. Requires `manage:all` (Super Admin) permission.",
        "operationId": "createUser",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "first_name",
                  "last_name",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "jane.doe@example.com"
                  },
                  "first_name": {
                    "type": "string",
                    "example": "Jane"
                  },
                  "last_name": {
                    "type": "string",
                    "example": "Doe"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "securepassword123"
                  },
                  "roleId": {
                    "type": "string",
                    "description": "Optional role ID to assign to the user.",
                    "example": "clx1y2z3a0000b4d2"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/v1/users/profile": {
      "get": {
        "summary": "Get current user profile",
        "description": "Returns the profile of the currently authenticated user.",
        "operationId": "getProfile",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user's profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "summary": "Update current user profile",
        "description": "Updates the email, first name, or last name of the currently authenticated user. Disabled in demo mode.",
        "operationId": "updateProfile",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "first_name": {
                    "type": "string"
                  },
                  "last_name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Disabled in demo mode.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/profile/password": {
      "post": {
        "summary": "Update password",
        "description": "Updates the password of the currently authenticated user. The current password must be provided for verification. Disabled in demo mode.",
        "operationId": "updatePassword",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "currentPassword",
                  "newPassword"
                ],
                "properties": {
                  "currentPassword": {
                    "type": "string",
                    "format": "password"
                  },
                  "newPassword": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Current password is incorrect.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Disabled in demo mode.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/{id}": {
      "get": {
        "summary": "Get a user",
        "description": "Returns a single user by ID. Requires `read:users` permission.",
        "operationId": "getUser",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "clx1y2z3a0000b4d2"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "summary": "Update a user",
        "description": "Updates a user's email, name, or role assignment. Requires `manage:all` (Super Admin) permission.",
        "operationId": "updateUser",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "clx1y2z3a0000b4d2"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "first_name": {
                    "type": "string"
                  },
                  "last_name": {
                    "type": "string"
                  },
                  "roleId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "summary": "Delete a user",
        "description": "Permanently deletes a user. Cannot delete the last remaining user. Requires `manage:all` (Super Admin) permission.",
        "operationId": "deleteUser",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "clx1y2z3a0000b4d2"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted. No content returned."
          },
          "400": {
            "description": "Cannot delete the only remaining user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    }
  },
...
}